libtracker-data: Improve error messages
The error currently reports the rule that failed, that is not that
much useful. Instead, accumulate all literals and terminals we pass
through in the parsing stage for the latest error position, and report
them all back.

These are precisely the parseable tokens we expected but didn't find,
hopefully gives a better clue at what did we want there.

Also, improve the handling of garbage after the query string, now we
simply report that we missed '\0'.

Example of the errors now:
Parser error at byte 0, expected one of 'base', 'prefix', 'select', 'construct', 'describe', 'ask'