cli: Add `tracker export --type=` option
This allows exporting a subset of data, rather than the whole graph.
It's intended to allow exporting user data from Tracker 2.x so we
can migrate app data from Tracker 2.x to Tracker 3.

In the exported output, files are identified by their URL rather than
the URN, as the same file will have a different URN in the Tracker 3.0
database.  It will be up to the relevant import tools to resolve this.