libtracker-sparql: Fix possible invalid writes in TrackerResource dispose
A resource may indirectly point to itself, in which case, clearing the
properties hashtables on finalize() will end up calling dispose() again
on this resource, but the hashtables were now gone.

Ensure the hashtables are freed/nullified on dispose(), so this kind
of reentrancy cannot happen. Any later time(s) the dispose method happens
will be on empty hashtables.