libtracker-data: Release non-select statements
Commit ed204139a1a added this extra flag to fix a race condition with
select statements, but broke the accounting for update statements that
don't get a TrackerDBCursor. This made statements in the update cache
essentially non-reusable, making the update cache ineffective.

Fix this by releasing the update statement after execution. This relies
on an update statement only being executed once after fetching from the
update cache, but that's how things work already.

This makes update statements in the cache again able to be reused in the
future, avoiding needlessly having to compile new statements from scratch.