libtracker-data: Avoid unescaping unichar characters if unneeded
Looking for unicode characters to unescape turns out a bit expensive on
multi-megabyte updates. Have a faster path for straight query copies if
there's no escaping sequence to be seen.

This turned out to be a fair chunk of time spent on tracker-miner-fs
first index.

5.39%     5.34%  pool-tracker-mi  libtracker-data.so                 [.] tracker_unescape_unichars

Now almost gone as unicode escaping characters are actually not that
much likely to happen often.