libtracker-sparql: Plug GError leak in TrackerEndpointDBus
The error handling in handle_cursor_reply() is somewhat iffy, on
one hand it "takes" an error without freeing it, and on the other
hand it can reuse the error variable for the handling of inner
errors.

Make the error argument const (and freed outside the handle_cursor_reply
call) and use a separate GError for the inner write_cursor() handling.