libtracker-data: Use md5 to uniquify graph in-memory databases
If a database is opened in memory, new graphs would be created as
further in-memory databases, however there is a clash with our
default graph naming scheme. The "tracker:" prefix expands to
"https://tracker.api.gnome.org/...#" and further characters of
the file name after that '#' are ignored. This makes all our
"tracker:*" graphs resolve to the same in-memory database, causing
glitches and warnings.

Get a md5 out of the graph name, so it's unique enough, and free
of troublesome characters. This makes the different graphs resolve
to separate in-memory databases and everything work again.

Could be seen with eg:
$ tracker3 endpoint -b org.foo.Bar -o nepomuk
$ tracker3 sparql -b org.foo.Bar -u -q "CREATE GRAPH tracker:Yay"
$ tracker3 sparql -b org.foo.Bar -u -q "CREATE GRAPH tracker:Ouch"