libtracker-data: Rebuild TrackerSparql SQL string on graph changes
If graphs are added or removed, the resulting SQL may be either
optimized better, or be broken as per the current database schema
(eg. if graph being removed was specified somewhere in the query).

Use the data manager "generation" number (increased on every graph
change) to find out whether the statement needs rebuilding before
execution, and do so. This will generate again a SQL string that
is up-to-date with the current graphs.