build: Move trackertestutils to regular libdir
This models what gobject-introspection does with its giscanner library,
which is the model I think we should copy (it's a Python module but
we consider it Tracker-specific rather than part of the regular Python
module ecosystem).

Previously we tried to put it into lib/ rather than lib64/ on 64-bit
systems. This causes unneeded friction.