Sanitize usage of libtracker-sparql/tracker-sparql.h includes
While this header is convenient as it contains all that
libtracker-sparql has to provide, the file indirectly depends on
generated sources for the enum types.

We simply don't need those enum types beneath libtracker-sparql,
so use the specific includes we need. This fixes build errors with
parallel builds, as there is no explicity dependency from
libtracker-data to the generated enum sources (and it would be
completely backwards).