libtracker-data: Create graph databases with a separate interface
Due to the way we create graphs (either implicitly on updates, or
via a CREATE GRAPH update), those always happen inside a DB
transaction, whereas WAL is documented to be ineffective (immediately
at least) if set inside a transaction.

This leaves freshly created database connections running on the
"delete" journal mode, even though WAL is "set", and would eventually
apply the next time the database is opened. This journal mode has
many drawbacks (worse performance, readers and writers possibly stalling
each other, ...) that are noticeable on eg. first index.

To fix this, delegate database file construction to a separate writable
interface created exclusively for this task, the WAL journal mode and
page size will be set there, so it is ensured to open with the right
journal mode when we open it with the readwrite interface.

Immediate benefits:
- First time tracker-miner-fs indexing (not involving tracker-extract)
  went locally from ~15s to ~9s, much closer to 2.x performance.
- Select queries won't stall, hopefully fixes some of those weird timeouts
  we've been seeing.