libtracker-data: Show function name in internal SPARQL function errors
I was trying to debug a query problem, with only this error message:

        Invalid arguments

I had to look thru tracker.git  to figure out the cause.
With this patch, the error becomes:

        tracker:uri-is-parent: Invalid arguments

This gives an immediate clue where the problem is in my query. Even
though most of these errors indicate an internal problem in Tracker,
accurate errors will help use trace and fix them.