libtracker-sparql: Query notifier data in batches
We have limits with the services table, and other limits on the number
of parameters higher up. We shouldn't be doing queries with an arbitrary
number of those.

Limit the batch size in queries for extra TrackerNotifier data, and
adapt code to maybe do extra queries for a single event batch.

As it is now possible that we are initiating queries on a fresh new
TrackerNotifierEventCache when an old one is still being processed, and
this happens from different threads (the update thread for the initial
one triggered from commit hooks, the main context thread for later async
processing), protect the statements execution with a mutex, as binding
arguments for execution is not thread safe.