libtracker-data: Do not leak variables when using ValuesClause in select
If ValuesClause is used in the topmost select, we surround the generated
query with a "SELECT * FROM (...) NATURAL INNER JOIN $values_subselect",
this makes the "SELECT *" the topmost query, and leaks variables from the
right hand side that are not meant to be in the resultset.

Avoid doing this as a nested subselect if ValuesClause were in the topmost
select, so that we return the expected avariables.