libtracker-data: Check for variables in the most direct context in SERVICE{}
When joining the service graph pattern with prior triple data in the same
context, we are looking for variables to join that are possibly out of the
current scope.

This shouldn't be the case, and leads to SQL-level errors like:

  no such column: Left.v_s

Make it sure we are only joining variables in the scope of the service query,
which yields the correct results.

Fixes: https://gitlab.gnome.org/GNOME/tracker/-/issues/248