libtracker-data: Make a static library
There's no multiple executables pulling this library directly. We have
some direct uses of this library, but they are all tests and non-install
executables in the repo.

Making it a static library won't increase install size, as only
libtracker-sparql pulls from it in the system, that's what everyone should
use instead.