manpages: Update references to the tracker3 CLI tool
We did reference it in manpages as simply "tracker", which is definitely
not going to be the case. Make CLI examples, titles and section names
correctly reference the real "tracker3" name.

Note: The name change also changes the target manpage file written by
our xsltproc step, since it is defined by the file+xsl. This will silently
go ignored, as installation of manpages is broken (we install empty files).
This will get fixed in later commits.