manpages: Fix installation of manpage files
The xsltproc step has two small gotchas: the output does not happen on
stdout, so we are capturing empty stdout, and installing an empty file instead
of our nice manpages.

But also, the target filename gets determined by the XSL based on the input,
so a manpage filename is as defined in its NAME section. This means we cannot
do filename changes to add the version this late in the stage.

Furthermore, the tracker-xdg-portal-3 manpage follows a different naming scheme
(i.e. not "tracker3-*").

Handle this all at once, specify per manpage target files that do match the
xsltproc output, and ensure the file gets generated at the current build dir.