libtracker-data: Lighten up TrackerSparql struct
There is an internal struct carrying the parser state, this struct began
small, but kept growing. This state is actually unneeded after parsing, so
we can keep TrackerSparql smaller, and only allocate the state transiently
during parsing.

Since TrackerSparql is the basis for TrackerSparqlStatement, this should
help memory usage a tiny bit if many prepared statements are cached.