libtracker-data: Fix tracker:title-sort
With a hack really, up in the parse tree we do force "COLLATE TRACKER"
on all string arguments passed on ORDER BY. Make this return "unknown"
just so this is avoided.

tracker:title-sort() is not meant to produce a pretty visible string,
more a collation key, perhaps it is somewhat fitting, perhaps I'm
making up an excuse.