Move file-utils from libtracker-common into libtracker-data
We only use these in one place, so we can move them there.

This should fix https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/140
in which the file-utils symbols conflict with those defined in
libtracker-miners-common.

The unit tests are gone, I'm not sure how to avoid that since we need
to avoid exporting these functions.