libtracker-data: Unset graph after Modify clause
If the Modify clause sets a graph via 'WITH' the graph token remains
but the solution underneath gets freed. That means the graph token's
literal is kept "set" but containing garbage, and may wreak havoc if
there are more Update clauses in the same update string that don't
specify a graph, as that garbage value will be used. This may lead
to warnings, crashes, and bogus graph databases being created. Eek.

The 'WITH' graph is not meant to be carried over, so ensure it gets
unset after dealing with the Modify clause.