portal: Fix initialization order
The portal is currently first claiming the bus name, then adding
the portal object+interface. This may break things with autostart
as clients are able to send a message to an object path that is not
there yet.

Changing the order means the object path is there when the DBus name
is made known, so clients are able to talk immediately to it.

Fixes the error reported at
https://github.com/flathub/org.gnome.Music/pull/24#issuecomment-702565846