libtracker-data: Use correct binding types for datetimes
When creating a binding from a literal datetime value, we mistakenly
use doubles (as used to be the common thing in Tracker 2.x). We now
store datetimes as either iso8601 strings or int64 timestamps, so
forward these types.

This code is in consistence with the update bits at tracker-data-update.c.

Fixes select queries like:
SELECT ("2020-01-01T01:01:01Z"^^xsd:dateTime AS ?date) { }