libtracker-data: Break out of all loops on transaction errors
If an error is found when flushing a transaction on a specific, we'd
inadvertently still try to handle operations in other graphs, possibly
reusing the GError location, and leading to invalid reads/writes.

After finding an error, the transaction is going to be rolled back
anyway, so break on the first error found.

Fixes: https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/130