libtracker-sparql: Batch TrackerNotifier queries
If we happen to get a high enough amount of updates, batched changes
accumulate, all querying for the additional information. This
accumulation of the same query leads at best to compiling new
statements (as the cached one is already in use). At worst, it leads
to the creation of a new TrackerDBInterface, with the initialization
cost involved.

Serializing this work, we're more likely to reuse a TrackerDBInterface
and a pre-compiled statement from previous runs.