libtracker-data: Favor free interfaces over new ones harder
We typically do that, but once we've found that the first interface
in the pool is already busy with other statements, we assume they're
all busy and jump into creating a new interface. Even worse, if that
interface is kept busy indefinitely, every new query will opt for
creating a new interface, quickly filling the allotment.

There might be other interfaces in the pool that are already
free, so check all of them before trying to create a new interface.
This makes us more conservative at creating new interfaces on load
periods, only doing so if all interfaces are actually busy.