libtracker-data: Inverse expectations when inserting a resource
Before this change we would first query the ID, and then insert it if non
existent. Apply the "better to ask for forgiveness than permission"
principle here, and try to always insert the resource. Only resort to
queries if there is a constraint violation (meaning the URI already
exists).

This optimizes insertions over lookups of existing resource IDs, but we
have other caching layers for the latter.