libtracker-data: Throw away the ontology/regular resource ID distinction
This only mattered for journals and replays. We don't have that anymore.
Furthermore, the split pool of IDs means we don't have one overflow
condition but two, and we handle neither well.

Letting the ID pool be shared means we may let SQLite pick a ROWID for us,
and it will definitely fare better if ROWIDs flipped the counter. It would
still result in an unused ROWID, we would badly break.