libtracker-data: Add fast path to insert new bnode UUIDs
The way to insert bnodes used to consist of
1) SELECT SparqlBNode() (which may query multiple times for an unused
   one).
2) SELECT ID FROM Resource WHERE Uri=bnode
3) INSERT INTO Resource(..., bnode)

Step 2 no longer applies since earlier commits, but this can be
streamlined so we try to insert bnodes until one comes up successful,
and it'll usually be on the first try.

Make this apply to direct updates from TrackerResource, as that's
where it's easiest to pull off.