libtracker-data: Cache set of new IDs in the transaction
Despite adding information for the same resource across several
graphs, we can find out that this is a new resource and go through
the fast paths that don't query prior info.

This makes adding information for a same resource across several
graphs cheaper, since we don't query for information we know beforehand
is not there.