libtracker-data: Do not implicitly delete rdf:type relation
Delete this explicitly before deleting the row, we want things explicit
as we'll be managing refcount in code.

There is also the slight inconsistence that deleting a "rdfs:Resource"
will have the rdf:type deletion go together with the handling of all
other rdfs:Resource properties, and then again on each of the subclasses
being deleted. Ensure this is done in one place, so refcount does not
get messed up.