libtracker-data: Handle variables for SERVICE clauses correctly
The sparql1.1 spec allows the use of variables in SERVICE clauses,
e.g.:

  ... SERVICE ?s { ... } ...

Where ?s is a variable resolved in other parts of the graph pattern.
https://www.w3.org/TR/2013/REC-sparql11-federated-query-20130321/#variableService
precisely advocates that query execution order should resolve the
possible values for the queried services before attempting to access
those.

Let Tracker do this somersault, if the service is a variable instead
of a literal propagate it up as a variable in the subselect so it is
matched together with all other variables when JOINed with the other
bits of the graph pattern, and let the services virtual table enforce
the execution order that resolves the service variable by returning
SQLITE_CONSTRAINT if the service column is missing. This makes SQLite
try the alternative execution orders, and eventually find one that
resolves the service variable when executing the remote service(s).

This allows delegating service queries to a variable, which may be
useful if e.g. we want to propagate a same query to multiple services:

  SELECT ?u {
    SERVICE ?s {
      ?u a nfo:Document
    }
  }
  VALUES ?s { 'dbus:...', 'dbus:...' }