libtracker-sparql: Allow prefixed names as per spec in TrackerResource
In order to figure out whether TrackerResource is dealing with a prefixed
name, we used g_uri_parse_scheme(). This happens to work for the most
common chars used in prefixes, however there's a substantial difference in
the charset allowed. For URI schemes (from
https://tools.ietf.org/html/rfc3986):

  scheme      = ALPHA *( ALPHA / DIGIT / "+" / "-" / "." )

For prefixed name prefixes (from
https://www.w3.org/TR/sparql11-query/#sparqlGrammar):

  PNAME_NS	  ::=  	PN_PREFIX? ':'
  PN_PREFIX	  ::=  	PN_CHARS_BASE ((PN_CHARS|'.')* PN_CHARS)?
  PN_CHARS_BASE	  ::=  	[A-Z] | [a-z] | [#x00C0-#x00D6] | [#x00D8-#x00F6] |
                        [#x00F8-#x02FF] | [#x0370-#x037D] | [#x037F-#x1FFF] |
                        [#x200C-#x200D] | [#x2070-#x218F] | [#x2C00-#x2FEF] |
                        [#x3001-#xD7FF] | [#xF900-#xFDCF] | [#xFDF0-#xFFFD] |
                        [#x10000-#xEFFFF]
  PN_CHARS_U	  ::=  	PN_CHARS_BASE | '_'
  PN_CHARS	  ::=  	PN_CHARS_U | '-' | [0-9] | #x00B7 | [#x0300-#x036F] |
                        [#x203F-#x2040]

Even though it's a bit of a layering break, lean on the SPARQL parser for
parsing prefixes exactly as per the spec.

Fixes: https://gitlab.gnome.org/GNOME/tracker/-/issues/286