libtracker-data: Check BIND variable in the current context
As per https://www.w3.org/TR/sparql11-query/#bind:
"The variable introduced by the BIND clause must not have been
used in the group graph pattern up to the point of use in BIND."

We are currently checking that the variable has bindings, but
that's also true if the variable was used in other graph patterns
prior to the current one.

We must check here only in the current context to match that
behavior. This makes queries like:

  SELECT ?u { { BIND (1 AS ?u) } UNION { BIND (2 AS ?u) } }

work as expected.