libtracker-data: Detect remote queries with no projected parameters
Querying a service that does not use any variable in its graph pattern
is fairly pointless. There are no variables to project into the rest
of the query, thus there's no way to match anything with the service
graph pattern. In pure SPARQL, that is basically just:

  SELECT * { SERVICE <...> { } }

But since in Tracker we are more lenient wrt giving each projected
variable a name, we also have cases like:

  SELECT * { SERVICE <...> { SELECT 2 { } } }

That are anyways pointless to execute. Detect all these cases and
send an empty query string instead, the service vtable will behave
as expected.