libtracker-sparql: Add tracker_sparql_connection_bus_new_async/finish
The bus connection started creating a dbus proxy that we did not need
to autostart, nor read properties from, so we could initially afford
to not have an asynchronous version.

But later we added more things on it, like session registration in the
portal, and pings to check for direct access permissions within a
possible sandbox.

It now makes sense to have that async variant, so add one.