libtracker-sparql: Fix race condition in tracker_sparql_connection_bus_new()
When commit cad0aefa0b added tracker_sparql_connection_bus_new_async(), it
made the sync variant run over the async variant for simplicity, using a
nested main loop that takes control until the asynchronous method returns.

However, we were using either of the thread default context, or the main
default one, in case of this call happening on a separate thread, this main
context may be iterating somewhere else, so the asynchronous method might
return before we've blocked on the main loop.

We want to lock the bus connection creation on this thread, so create a
new main context that is temporarily made the thread default, this will
ensure the asynchronous call happens in our own thread, under our own
main loop.