libtracker-data: Add paranoia check in condition
According to Coverity, there's weird paths that may cause an ontology
change on a new class that may still result on us trying to update the
range of a related property.

It remains to be seen whether the combination of circumstances that
lead to this can ever happen together, but it's simple enought to add
the paranoia check as those variables where we are composing the ALTER
TABLE would be actually empty.

CID: #1501144