libtracker-sparql, tests: Support empty array updates
This is for the convenience of users of the API where the length of
the array depends on runtime conditions and it's possible that it will
sometimes be empty.

The 'bus' and 'direct' implementations of the update_array_async
virtual method are mainly predicated on the length of the array. So, as
long as an empty array is accompanied by a zero length, it should work
as a NOP.

For the 'bus' implementation, another alternative was to handle it in
the endpoint (or the service-side) which would otherwise receive an
empty GUnixFDList. However, it's slightly nicer to handle it in the
client to avoid the IPC.

https://gitlab.gnome.org/GNOME/tracker/-/merge_requests/392