libtracker-data: Handle relational expressions with arbitrary datetimes
This had some holes, since we can store datetimes in 2 different formats
(unix timestamps and iso8601 datetime strings), but rely on direct comparisons
for all data types inside FILTER().

Detect date/time comparisons, and use the resulting SparqlTimeSort() to
compare the values on both sides of the relational expression, this takes
care of evening out the format and timezone differences so we can compare
all timestamps.