Make tracker:endpoint test work on macOS
The symptom on my system are that the test fails to run because gi
cannot find the shared library object for Tracker. This is due to the
fact that $LD_LIBRARY_PATH is set by Meson but removed from the
environment by the time the test is running. This is because the
examples/python/endpoint.py scripts which uses '!#/usr/bin/env python'
as a shebang and on macOS the env utility removes dynamic linker
variables from the environment. Fix this by explicitly using python to
execute the script.

Use this opportunity for restructure how the Python installation and
executable is looked up in the build definition: use the 'python'
Meson module to lock for a suitable python in the main meson.build.

Also, the test setup does not set $TRACKER_LANGUAGE_STOP_WORDS_DIR to
point to the stop words directory in the source directory. Fix it.