tracker-data-manager: Simplify tracker_data_manager_recreate_indexes()
This function first drops all indexes and then recreates them. For how
the code is structured, this forces all functions that modify the
indexes to have two modes: one mode in which the indexes are only
dropped, and another in which they are dropped and then recreated.

Recreating the indexes in two steps does not provide any apparent
advantage (the current code actually executes all DROP INDEX SQL
statements twice, thus it may be a bit slower overall).

Simplify the code and drop and recreate the indexes one at a time.
This will enable furhter simplifications.