libtracker-sparql: Solve signed vs unsigned pointer warnings
libxml2 functions 'const xmlChar *' (aka 'unsigned char *') type for
strings arguments however literals in C are of type 'char *'.  This
results in warnings about the conversion between pointers of different
size. Add a macro to perform the type casts required to silence the
warnings in a more descriptive and less ugly way.