libtracker-data: Only issue FTS updates where we mean to
Right now, getting the old values for a FTS property means those
values are deleted, and a FTS update is issued afterwards (even if
with empty strings).

We don't strictly want to do that from all paths, only from those
that are changing (i.e. not deleting) the FTS property value. Since
the FTS5 machinery is getting extremely picky, it is better to
avoid these needless roundtrips.