docs: Rewrite data model of documentation tool
Instead of parsing ontology TTL by itself, use 2 in-memory
TrackerSparqlConnections for the task, one to load .description
files, and another with an empty database.

With these in place, the ontology is fully introspectable, we
can then query these to fill in our information about the defined
classes and properties.

One advantage here is that we avoid purpose-specific Turtle
file parsers and unify this on the sparql library.