ci: Check code style in merge request changes
Use the new check-style script to check the code style for all
changes in a merge request, in order to suggest changes to conform
to it.

However, allow these checks to fail, so they are more of a hint
while the uncrustify configuration is tested to be correct, we
decide to enforce the style, etc.

As the script requires uncrustify, ensure it's installed in the
Fedora image, and use it to run the check.