add tests for feature of supporting comments in ontology files
There are no special tests that should be carried out to test the feature of skipping comments in ontology file.
Just we need to use any ontology file that contains different types of comments and make sure that the ontology is parsed well.
That is done by applying different queries on the database created and make sure that the queries return the right data which means that the schema is parsed successfully.