tracker-turtle-reader: add support for comments inside triple patterns
Fixes the crash that happens when the ontology file contains comments at the middle a triple pattern.
Before this commit, tracker only supports comments at the beginning or end of the triple pattern (either before the subject or after the dot).
This commit consider comments the same as white spaces, can appear anywhere in the ontology file outside the string or IRIREF.
So, I renamed skip_comments() function to advance_whitespaces_and_comments() and made it call advance_whitespaces() before checking for the '#' comment delimiter char.
advance_whitespaces_and_comments() by default assumes the comment be multilined, so it can skip and combination of comments and whitespaces.
Then, I replaced each call to the skip_whitespaces() by a call to advance_whitespaces_and_comments()

tmp