tracker-turtle-reader.c: track current line and column numbers
Add the functionality of tracking the current line and column numbers of the turtle file being parsed.
These numbers will be used to specify the location of the parsing warning or error when happen.
The seek_input() function is added, so it can be used to advance in the ontology file.
Each time a token is parsed, the seek_input() function is called to skip the parsed token.
It counts the number of characters and new lines in the parsed token and it update the current line and column numbers according to them.