tests: Disable GVFS in functional tests
On some of these tests, we create a per-testcase dbus environment,
thus rely on g_bus_get*() default connections to be correctly tear
down and created again for the next testcase.

GVFS does not seem to play along with this, and keeps the default
GDBusConnections alive between runs. This may fail with "Connection
already closed" if we attempted to get a default session/system bus
on later testcases, since the dbus daemon is actually tear down.

We actually don't need to pull GVFS here at all, so disable it
for our testcases to tiptoe the issue.

Fixes: https://gitlab.gnome.org/GNOME/tracker/-/issues/314