libtracker-data: Ensure to synchronize cached values after delete
In these paths of TrackerResource insertion handling for single valued
properties, we schedule property deletion, but don't synchronize the
cache here. Clear the value array contents so future checks match.

This is the most direct cause of the warning seen at
https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/179, there
is a double CREATED event emission going on there, that due to
non-atomicity in the file creation operation has those TrackerResources
disagreeing in the nfo:fileSize for the file.

Being those 2 tracker_resource_set_*() operations, we schedule prior
deletes, but at the time of handling the second nfo:fileSize we'd
still be thinking there was a value left in the array, due to the lack
of sync.

Fixes: https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/179